secure-sensor-data-gateway/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.31
marshmallow==3.21.3
python-dotenv==1.0.1
pytest==8.2.2
